+++ src/refactoring/components/admin/coupon/CouponList.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CouponList = () => {
+  return <div>CouponList</div>;
+};
+
+export default CouponList;
